% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcActualDisp.r
\name{calcActualDisp}
\alias{calcActualDisp}
\title{calcActualDisp}
\usage{
calcActualDisp(Cb, Cbw, actualDraft, maxDraft, maxDisplacement)
}
\arguments{
\item{Cb}{Maximum block coefficient (vector of numericals, dimensionless) (see \code{\link{calcCb}})}

\item{Cbw}{Waterline block coefficient (vector of numericals, dimensionless) (see \code{\link{calcCbw}})}

\item{actualDraft}{Actual draft (vector of numericals, m)}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{maxDisplacement}{Maximum ship displacement (vector of numericals, m^3)}
}
\value{
\code{actualDisplacement} (vector of numericals, m^3)
}
\description{
Estimate actual loaded displacement (\code{actualDisplacement})
(m^3) using a vessel's actual draft.
}
\details{
Loaded displacement is estimated using actual draft. Actual draft is
typically obtained from sources such as AIS messages or ship records.
Uses Riddlesworth method (MAN, 2011).
}
\examples{
calcActualDisp(c(0.82,0.66), c(0.81,0.65), c(12.5,11.1), c(13.6,11.5), c(80097,52382))
calcActualDisp(0.82, 0.81, 12.5, 13.6, 80097)

}
\references{
\href{https://www.man-es.com/marine/products/propeller-aft-ship}{MAN Energy
 Solutions. 2011. "Basic Principles of Propulsion." pg. 9.}
}
\seealso{
\code{\link{calcCb}}, \code{\link{calcCbw}}
}
