% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCf.r
\name{calcCf}
\alias{calcCf}
\title{calcCf}
\usage{
calcCf(shipSpeed, lwl, seawaterTemp = 15, seawaterDensity = 1.025)
}
\arguments{
\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see
\code{\link{calcSpeedUnitConversion}})}

\item{lwl}{Waterline length (vector of numericals, m) (see 
\code{\link{calclwl}})}

\item{seawaterTemp}{Sea water temperature. Default = 15 (degrees Celsius).
Can supply either a vector of numericals, a single number, or rely on the 
default}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can 
supply either a vector of numericals, a single number, or rely on the default}
}
\value{
\code{Cf} (vector of numericals, dimensionless)
}
\description{
Calculates the frictional resistance coefficient (\code{Cf})
(dimensionless).
}
\details{
Calculates frictional resistance coefficient using the International Tank
Towing Committee (ITTC) 1957 method where:
\deqn{Cf = \frac{0.075}{(log(V*Lwl/visc)-2)^2}}{Cf = 0.075/((log(V*Lwl/visc)-2)^2)}
}
\examples{
calcCf(seq(1,5,1), 218.75, 15, 1.025)

}
\references{
Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion
Power of Ships."
}
\seealso{
\itemize{
\item \code{\link{calcSpeedUnitConversion}}
\item \code{\link{calclwl}}
}

Other Resistance Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Resistance Calculations}
