% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCp.r
\name{calcCp}
\alias{calcCp}
\title{calcCp}
\usage{
calcCp(
  Cm,
  Cbw,
  shipType,
  bounds = "none",
  roroPaxContainerShipTypes = c("ro.ro", "passenger", "ferry.pax", "ferry.ro.pax",
    "cruise", "cruise.ed", "yacht", "container.ship"),
  gCargoShipTypes = c("general.cargo"),
  tankerBulkCarrierShipTypes = c("tanker", "chemical.tanker", "liquified.gas.tanker",
    "oil.tanker", "other.tanker", "bulk.carrier")
)
}
\arguments{
\item{Cm}{Midship section coefficient (vector of numericals, dimensionless) 
(see \code{\link{calcCm}})}

\item{Cbw}{Waterline block coefficient (vector of numericals, dimensionless) 
(see \code{\link{calcCbw}})}

\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}). 
Must align with \code{roroPaxContainerShipTypes},
\code{gCargoShipTypes}, and \code{tankerBulkCarrierShipTypes} groupings}

\item{bounds}{Indicates which upper and lower bounds on \code{Cp} should be
applied:\itemize{
\item Pass "holtrop mennen" to use the bounds specified by Holtrop & Mennen
\item Pass "none" (default) to calculate \code{Cp} without upper or lower
bounds}
This argument is not vectorized. Either supply a single string or rely on the
default}

\item{roroPaxContainerShipTypes}{Ship types specified in input \code{shipTypes}
to be modeled as RORO, passenger and container ships (vector of strings)}

\item{gCargoShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as general cargo (vector of strings)}

\item{tankerBulkCarrierShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers and bulk carriers (vector of strings)}
}
\value{
\code{Cp} (vector of numericals, dimensionless)
}
\description{
Calculates the prismatic coefficient (\code{Cp}) (dimensionless).
}
\details{
\deqn{Cp = \frac{Cbw}{Cm}}{Cp=Cbw/Cm}

This function can calculate \code{Cp} with or without upper and lower bounds.
If the Holtrop & Mennen bounds are applied, this requires ship types to be
grouped. Use the \code{roroPaxContainerShipTypes}, \code{gCargoShipTypes},
and  \code{tankerBulkCarrierShipTypes} parameters to provide these ship
type groupings. Any ship types not included in these groupings will be
considered as miscellaneous vessels.
}
\examples{
calcCp(c(0.99,0.98), c(.8,.75),c("bulk.carrier","container.ship"),"none")

}
\references{
\href{https://www.man-es.com/marine/products/propeller-aft-ship}{MAN Energy
Solutions. 2011. "Basic Principles of Propulsion."}
}
\seealso{
\itemize{
  \item \code{\link{calcCm}}
  \item \code{\link{calcCbw}}
  \item \code{\link{calcShipType}}
}
}
