% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDispUnitConversion.r
\name{calcDispUnitConversion}
\alias{calcDispUnitConversion}
\title{calcDispUnitConversion}
\usage{
calcDispUnitConversion(tonnageDisp, seawaterDensity = 1.025)
}
\arguments{
\item{tonnageDisp}{Displacement (vector of numericals, tonnage)}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can 
supply either a vector of numericals, a single number, or rely on the default}
}
\value{
\code{disp} (vector of numericals, m^3)
}
\description{
Convert displacement in tonnages to cubic meters.
}
\examples{
calcDispUnitConversion(70000, 1.025)

}
