% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFroudeNum.r
\name{calcFroudeNum}
\alias{calcFroudeNum}
\title{calcFroudeNum}
\usage{
calcFroudeNum(shipSpeed, lwl)
}
\arguments{
\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see 
\code{\link{calcSpeedUnitConversion}})}

\item{lwl}{Waterline length (vector of numericals, m) (see 
\code{\link{calclwl}})}
}
\value{
\code{froudeNum} (vector of numericals, dimensionless)
}
\description{
Calculate the Froude number (\code{froudeNum}) (dimensionless).
}
\details{
The Froude number is a dimensionless coefficient that relates a vessel's speed
and length to a relative resistance. The equation for the Froude number is:
\deqn{Fn=\frac{V}{\sqrt{g*Lwl}}}{Fn=V/sqrt(g*lwl)}
Where V is the ship's speed, g is the gravitational constant (9.81 m/s^2),
and \code{lwl} is the waterline length (m) of the vessel.
}
\examples{
calcFroudeNum(seq(1,5,1),214)

}
\references{
\href{https://www.man-es.com/marine/products/propeller-aft-ship}{MAN Energy
Solutions. 2011. "Basic Principles of Propulsion."}
}
\seealso{
\itemize{
\item \code{\link{calcSpeedUnitConversion}}
\item \code{\link{calclwl}}
}
}
