% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMAppendageRes.r
\name{calcHMAppendageRes}
\alias{calcHMAppendageRes}
\title{calcHMAppendageRes}
\usage{
calcHMAppendageRes(
  shipSpeed,
  Cf,
  appendagesList,
  wettedAppSAList,
  seawaterDensity = 1.025
)
}
\arguments{
\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see 
\code{\link{calcSpeedUnitConversion}})}

\item{Cf}{Frictional resistance coefficient (vector of numericals, 
dimensionless) (see \code{\link{calcCf}})}

\item{appendagesList}{List of appendages on ship (vector of strings) \itemize{
\item"rudder behind skeg"
\item"rudder behind stern"
\item"twin-screw balance rudders"
\item"shaft brackets"
\item"skeg"
\item"strut bossings"
\item"hull bossings"
\item"shafts"
\item"stabilizer fins"
\item"dome"
\item"bilge keels"
}}

\item{wettedAppSAList}{List of wetted surface areas corresponding to list of appendages
(vector of numericals, m^2)}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can 
supply either a vector of numericals corresponding to the ship speed, provide a single
number, or rely on the default}
}
\value{
\code{Rapp} (vector of numericals, kN)
}
\description{
Calculate appendage resistance (\code{Rapp}) (kN) from the
Holtrop & Mennen method.
}
\examples{
calcHMAppendageRes(seq(1,5,1),0.0015,"rudder behind skeg",50,seawaterDensity=1.025)
calcHMAppendageRes(seq(1,5,1),0.0015,NA,0,seawaterDensity=1.025)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.
}
\seealso{
\itemize{
\item \code{\link{calcSpeedUnitConversion}}
\item \code{\link{calcCf}} }

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Holtrop-Mennen Calculations}
\concept{Resistance Calculations}
