% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMBulbousBowRes.r
\name{calcHMBulbousBowRes}
\alias{calcHMBulbousBowRes}
\title{calcHMBulbousBowRes}
\usage{
calcHMBulbousBowRes(
  shipSpeed,
  maxDraft,
  forwardDraft,
  Abt,
  hb,
  seawaterDensity = 1.025
)
}
\arguments{
\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see 
\code{\link{calcSpeedUnitConversion}})}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{forwardDraft}{Forward draft (deviation from actual draft indicates trim) 
(vector of numericals, m)}

\item{Abt}{Traverse bulb area (vector of numericals, m^2) (see 
\code{\link{calcAbt}})}

\item{hb}{Center of bulb area above keel line (vector of numericals, m) 
(see \code{\link{calchb}})}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can 
supply either a vector of numericals, a single number, or rely on the default}
}
\value{
\code{Rb} (vector of numericals, kN)
}
\description{
Calculate bulbous bow resistance (\code{Rb}) (kN) from the
Holtrop & Mennen method.
}
\examples{
calcHMBulbousBowRes(seq(1,5,1),13.57,13.57,25.09,5.43,seawaterDensity=1.025)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.
}
\seealso{
\itemize{
\item \code{\link{calcSpeedUnitConversion}}
\item \code{\link{calcAbt}}
\item \code{\link{calchb}}
}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Holtrop-Mennen Calculations}
\concept{Resistance Calculations}
