% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMCa.r
\name{calcHMCa}
\alias{calcHMCa}
\title{calcHMCa}
\usage{
calcHMCa(maxDraft, lwl, Cbw, breadth, forwardDraft, Abt, hb)
}
\arguments{
\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{lwl}{Waterline length (vector of numericals, m) (see \code{\link{calclwl}})}

\item{Cbw}{Waterline block coefficient (vector of numericals, dimensionless) 
(see \code{\link{calcCbw}})}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{forwardDraft}{Forward draft (deviation from actual draft indicates trim)
(vector of numericals, m)}

\item{Abt}{Traverse bulb area (vector of numericals, m^2) (see 
\code{\link{calcAbt}})}

\item{hb}{Center of bulb area above keel line (vector of numericals, m) (see 
\code{\link{calchb}})}
}
\value{
\code{Ca} (vector of numericals, dimensionless)
}
\description{
Calculate incremental hull (roughness) resistance coefficient
(\code{Ca}) (dimensionless), which describes the effect of hull roughness and
still-air resistance, using the Holtrop & Mennen method.
}
\examples{
calcHMCa(c(13.57,11.49),c(218.7500,209.2518),c(0.81,0.65),c(32.25,32.20),c(13.57,11.49),
c(25.09,55.86),c(5.43,4.6))

calcHMCa(13.57,218.7500,0.81,32.25,13.57,25.0880,5.428)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.
}
\seealso{
\itemize{
\item \code{\link{calclwl}}
\item \code{\link{calcCbw}}
\item \code{\link{calcAbt}}
\item \code{\link{calchb}}
}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Holtrop-Mennen Calculations}
\concept{Resistance Calculations}
