% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMFormFactor.r
\name{calcHMFormFactor}
\alias{calcHMFormFactor}
\title{calcHMFormFactor}
\usage{
calcHMFormFactor(
  maxDraft,
  lwl,
  breadth,
  maxDisplacement,
  Cp,
  Cstern = 0,
  lcb = 0
)
}
\arguments{
\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{lwl}{Waterline length (vector of numericals, m) (see \code{\link{calclwl}})}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{maxDisplacement}{Maximum ship displacement (vector of numericals, m^3)}

\item{Cp}{Prismatic coefficient (vector of numericals, dimensionless) (see 
\code{\link{calcCp}})}

\item{Cstern}{Afterbody form coefficient:
\itemize{\item V-shaped Hull = -10
        \item U-Shaped Hull =  10
        \item Normal Hull = 0 (default) }
Can supply either a vector of numericals, a single number, or rely on the default}

\item{lcb}{Longitudinal position of center of buoyancy (vector of numericals,
see \code{\link{calclcb}})}
}
\value{
\code{formFactor} (vector of numericals)
}
\description{
Calculate form factor (1+k_1) from the Holtrop & Mennen method.
}
\examples{
calcHMFormFactor(c(13.57,11.49),c(218.75, 209.25),c(32.25,32.20),c(80097,52382.04),c(0.81,0.67))
calcHMFormFactor(13.57,218.75,32.25,80097,0.81)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.

Holtrop, J. and Mennen, G. G. J. 1984. "A Statistical Re-Analysis of Resistance
and Propulsion Data'.
}
\seealso{
\itemize{
\item \code{\link{calclwl}}
\item \code{\link{calcCp}}
\item \code{\link{calclcb}} }

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}
}
\concept{Holtrop-Mennen Calculations}
