% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMImmersedTransomRes.r
\name{calcHMImmersedTransomRes}
\alias{calcHMImmersedTransomRes}
\title{calcHMImmersedTransomRes}
\usage{
calcHMImmersedTransomRes(
  shipSpeed,
  breadth,
  Cwp,
  maxDraft,
  At,
  seawaterDensity = 1.025
)
}
\arguments{
\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see
\code{\link{calcSpeedUnitConversion}})}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{Cwp}{Water plane area coefficient (vector of numericals, see 
\code{\link{calcCwp}})}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{At}{Transom area (vector of numericals, m^2) (see \code{\link{calcAt}})}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can 
supply either a vector of numericals, a single number, or rely on the default}
}
\value{
\code{Rtr} (vector of numericals, kN)
}
\description{
Calculate additional pressure resistance due to immersed transom (\code{Rtr})
(kN).
}
\examples{
calcHMImmersedTransomRes(seq(1,5,1),32.25,0.91,13.57,22.2,seawaterDensity=1.025)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.
}
\seealso{
\itemize{
\item \code{\link{calcSpeedUnitConversion}}
\item \code{\link{calcCwp}}
\item \code{\link{calcAt}} }

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Holtrop-Mennen Calculations}
\concept{Resistance Calculations}
