% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMPwr.r
\name{calcHMPwr}
\alias{calcHMPwr}
\title{calcHMPwr}
\usage{
calcHMPwr(
  totalInstalledPwr,
  shipSpeed,
  actualDraft,
  maxDraft,
  shipType,
  lwl,
  breadth,
  maxDisplacement,
  Cb,
  nProp,
  serviceMargin = 15,
  shaftEff = 0.98,
  relRotationEff = 1,
  seawaterTemp = 15,
  seawaterDensity = 1.025,
  pwrUpperBoundPercent = 1,
  pwrLowerBoundPercent = 0.02,
  CmEquationType = "kristensen",
  tankerBulkCarrierShipTypes = c("tanker", "chemical.tanker", "liquified.gas.tanker",
    "oil.tanker", "other.tanker", "bulk.carrier"),
  tugShipTypes = c("service.tug", "tug"),
  roroPaxShipTypes = c("passenger", "ferry.pax", "ferry.ro.pax", "cruise", "cruise.ed",
    "yacht", "ro.ro"),
  gCargoShipTypes = c("general.cargo"),
  containerShipTypes = c("container.ship"),
  Cstern = 0,
  forwardDraft = NULL,
  aftDraft = NULL,
  appendagesList = c(""),
  wettedAppSAList = NA
)
}
\arguments{
\item{totalInstalledPwr}{Total installed main engine power (vector of numericals, kW) (maximum
continuous rated power)}

\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see 
\code{\link{calcSpeedUnitConversion}})}

\item{actualDraft}{Actual draft (vector of numericals, m)}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}). 
Must align with \code{tankerBulkCarrierShipTypes}, \code{tugShipTypes}, 
\code{roroPaxShipTypes}, \code{gCargoShipTypes},
\code{containerShipTypes} groupings}

\item{lwl}{Waterline length (vector of numericals, m) (see \code{\link{calclwl}})}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{maxDisplacement}{Maximum ship displacement (vector of numericals, m^3)}

\item{Cb}{Maximum block coefficient (vector of numericals, dimensionless) (see 
\code{\link{calcCb}})}

\item{nProp}{Number of propellers (vector of numericals, see \code{\link{calcPropNum}})}

\item{serviceMargin}{A service margin to account for weather and sea effects:
\itemize{\item At-sea operations = 15 (Default) \item Coastal operations = 10} Can 
supply either a vector of numericals, a single number, or rely on the default}

\item{shaftEff}{Shaft efficiency (dimensionless). Default = 0.98.
Ratio of power delivered to the propeller and the brake power delivered by
the engine. Can supply either a vector of numericals, a single number, or rely on the default.}

\item{relRotationEff}{Relative rotational efficiency (dimensionless).
 Default = 1. Accounts for effect of rotational flow of water around propeller.
Can supply either a vector of numericals, a single number, or rely on the default}

\item{seawaterTemp}{Sea water temperature. Default = 15 (degrees Celsius). Can 
supply either a vector of numericals, a single number, or rely on the default}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can 
supply either a vector of numericals, a single number, or rely on the default}

\item{pwrUpperBoundPercent}{Percent of total installed power at which
required power is capped. Default = 1, which indicates required power cannot
exceed \code{totalInstalledPwr}. Can supply either a vector of numericals, a 
single number, or rely on the default}

\item{pwrLowerBoundPercent}{Percent of total installed power to act as lower
bound for required power. Default = 0.02, which indicates required power
cannot go below 2\% of \code{totalInstalledPwr}. Can supply either a vector
of numericals, a single number, or rely on the default}

\item{CmEquationType}{Type of equation to estimate the midship section
coefficient (see \code{\link{calcCm}}): \itemize{
\item"kristensen (Default)"
\item"benford"
\item"schneekluth"}
This argument is not vectorized, as it takes only a single string}

\item{tankerBulkCarrierShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers and bulk carriers}

\item{tugShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as tugs (vector of strings)}

\item{roroPaxShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as RORO and passenger ships (vector of strings)}

\item{gCargoShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as general cargo (vector of strings)}

\item{containerShipTypes}{Ship types specified in input \code{shipTypes} to
be modeled as container ships (vector of strings)}

\item{Cstern}{Afterbody form coefficient: \itemize{\item V-shaped Hull = -10
        \item U-Shaped Hull =  10
        \item Normal Hull = 0 (default)}
Can supply either a vector of numericals, a single number, or rely on the default}

\item{forwardDraft}{Forward draft (deviation from actual draft indicates trim) 
(vector of numericals, m)}

\item{aftDraft}{Aft draft (deviation from actual draft indicates trim)
(vector of numericals, m)}

\item{appendagesList}{List of appendages on ship (vector of strings) \itemize{
\item"rudder behind skeg"
\item"rudder behind stern"
\item"twin-screw balance rudders"
\item"shaft brackets"
\item"skeg"
\item"strut bossings"
\item"hull bossings"
\item"shafts"
\item"stabilizer fins"
\item"dome"
\item"bilge keels"}}

\item{wettedAppSAList}{List of wetted surface areas corresponding to list of
appendages (vector of numericals, m^2)}
}
\value{
power (vector of numericals, kW)
}
\description{
Calculate instantaneous main engine power (kW) using the Holtrop & Mennen method.
}
\details{
Primary method from Holtrop & Mennen (1982). Updated equations for high speed
operations (Froude number > 0.55) from Holtrop & Mennen (1984). Estimation of
some inputs use methodology from Rakke (2016).

This method this requires ship types to be grouped. Use the
\code{tankerBulkCarrierShipTypes}, \code{tugShipTypes}, \code{roroPaxShipTypes},
\code{gCargoShipTypes}, \code{containerShipTypes} grouping parameters to
provide these ship type groupings. Any ship types not included in these groupings
will be considered as miscellaneous vessels.

Ship speed and actual draft are typically obtained from sources such as AIS
messages or ship records.
}
\examples{
calcHMPwr(
totalInstalledPwr=rep(9363,2),,
shipSpeed=seq(0,1,1),
actualDraft=rep(12.48,2),
maxDraft=rep(13.57,2),
shipType=rep("bulk.carrier",2),
lwl=rep(218.75,2),
breadth=rep(32,2),
maxDisplacement=rep(80097,2),
Cb=rep(0.8099003,2),
nProp=rep(1,2),
serviceMargin=15,
shaftEff=0.98,
relRotationEff=1,
seawaterTemp=15,
seawaterDensity=1.025,
Cstern=0,
CmEquationType="kristensen",
forwardDraft=rep(13.57,2),
aftDraft=rep(13.57,2),
appendagesList=c(""),
wettedAppSAList=NA
)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.

Holtrop, J. and Mennen, G. G. J. 1984. "A Statistical Re-Analysis of Resistance
and Propulsion Data'.

\href{http://hdl.handle.net/11250/2410741}{Rakke, S. G. 2016. "Ship Emissions
Calculation from AIS." NTNU.}
}
\seealso{
\itemize{
\item \code{\link{calcSpeedUnitConversion}}
\item \code{\link{calclwl}}
\item \code{\link{calcCb}}
\item \code{\link{calcPropNum}}
\item \code{\link{calcCm}}
\item \code{vignette("OverviewOfPowerModels", package="ShipPowerModel")}
\item \code{vignette("HoltropMennen.Example", package="ShipPowerModel")}
}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}
}
\concept{Holtrop-Mennen Calculations}
