% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMThrustFactor.r
\name{calcHMThrustFactor}
\alias{calcHMThrustFactor}
\title{calcHMThrustFactor}
\usage{
calcHMThrustFactor(
  breadth,
  lwl,
  maxDraft,
  maxDisplacement,
  nProp,
  Cp,
  propDiam,
  Cbw,
  Cstern = 0,
  lcb = 0,
  seawaterDensity = 1.025
)
}
\arguments{
\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{lwl}{Waterline length (vector of numericals, m) (see \code{\link{calclwl}})}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{maxDisplacement}{Maximum ship displacement (vector of numericals, m^3)}

\item{nProp}{Number of propellers (vector of numericals, see 
\code{\link{calcPropNum}})}

\item{Cp}{Prismatic coefficient (vector of numericals, dimensionless) 
(see \code{\link{calcCp}})}

\item{propDiam}{Propeller diameter (vector of numericals, m) (see 
\code{\link{calcPropDia}})}

\item{Cbw}{Waterline block coefficient (vector of numericals, dimensionless) (see
\code{\link{calcCbw}})}

\item{Cstern}{Afterbody form coefficient: \itemize{\item V-shaped Hull = -10
        \item U-Shaped Hull =  10
        \item Normal Hull = 0 (default)}
Can supply either a vector of numericals, a single number, or rely on the default}

\item{lcb}{Longitudinal position of center of buoyancy (vector of numericals, see 
\code{\link{calclcb}})}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can supply
either a vector of numericals, a single number, or rely on the default}
}
\value{
\code{thrustFactor} (vector of numericals, dimensionless)
}
\description{
Calculate thrust deduction factor (\code{thrustFactor})
(dimensionless) from the Holtrop & Mennen method.
}
\details{
Thrust deduction factor is a component of hull efficiency as well as a
component of propeller efficiency. It describes the increase in resistance on
the hull from water getting sucked back towards the propeller.

Note: In "A Statistical Re-Analysis of Resistance and Propulsion Data", the
authors re-analyze with the inclusion of Series 64 hull forms for a total of
334 models included in the analysis. They suggest an update of the single
screw thrust equation but that the original equations should be used
for twin screw ships.

Additionally, we are assuming conventional stern for all single screw ships.
Holtrop & Mennen also include an estimation for thrust factor for single screw
ships with open stern for fast sailing ships, but that is not included here.
}
\examples{
calcHMThrustFactor(c(32.25,32.20),c(218.75,209.25),c(13.57,11.49),c(80097,52382.04),
                   c(1,1),c(0.81,0.67),c(6.7,7),c(0.81,0.65))
calcHMThrustFactor(32.25,218.75,13.57,80097,
                   1,0.81,6.7,0.81)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.

Holtrop, J. and Mennen, G. G. J. 1984. "A Statistical Re-Analysis of Resistance
and Propulsion Data'.
}
\seealso{
\itemize{
\item \code{\link{calclwl}}
\item \code{\link{calcPropNum}}
\item \code{\link{calcCp}}
\item \code{\link{calcPropDia}}
\item \code{\link{calcCbw}}
\item \code{\link{calclcb}}}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}
}
\concept{Holtrop-Mennen Calculations}
