% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMTotalRes.r
\name{calcHMTotalRes}
\alias{calcHMTotalRes}
\title{calcHMTotalRes}
\usage{
calcHMTotalRes(
  Rapp,
  Rw,
  Rb,
  Rtr,
  seawaterDensity,
  wettedSA,
  shipSpeed,
  Cf,
  formFactor,
  Ca,
  serviceMargin
)
}
\arguments{
\item{Rapp}{Appendage resistance (vector of numericals, kN) (see 
\code{\link{calcHMAppendageRes}})}

\item{Rw}{Wave waking (similar to residual) resistance (vector of numericals, kN)
(see \code{\link{calcHMWaveMakingRes}})}

\item{Rb}{Bulbous bow resistance (vector of numericals, kN) (see 
\code{\link{calcHMBulbousBowRes}})}

\item{Rtr}{Immersed transom resistance (vector of numericals, kN) (see 
\code{\link{calcHMImmersedTransomRes}})}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^2). Can 
supply either a vector of numericals, a single number, or rely on the default}

\item{wettedSA}{Wetted hull surface area (vector of numericals, m^2) (see 
\code{\link{calcHMWettedSA}})}

\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see 
\code{\link{calcSpeedUnitConversion}})}

\item{Cf}{Frictional resistance coefficient (vector of numericals, dimensionless) 
(see \code{\link{calcCf}})}

\item{formFactor}{Form factor (1+k) (vector of numericals, dimensionless)
(see \code{\link{calcHMFormFactor}})}

\item{Ca}{Incremental hull (roughness) resistance coefficient (vector of 
numericals, dimensionless) (see \code{\link{calcHMCa}})}

\item{serviceMargin}{A service margin to account for weather and sea effects:
\itemize{\item Coastal operations = 10 \item At-sea operations = 15 (default)}.
Can supply either a vector of numericals, a single number, or rely on the default}
}
\value{
\code{Rtot} (vector of numericals, kN)
}
\description{
Calculate total ship resistance (\code{Rtot}) (kN) using the
Holtrop & Mennen method.
}
\details{
Note that service margin is included here as a resistance term.
}
\examples{
calcHMTotalRes(Rapp=0.24,
               Rw=2.6,
               Rb=c(4.538188e-07,3.579217e-06,1.180116e-05,2.709520e-05,5.085940e-05),
               Rtr=c(0.58,2.2,4.8,8.1,12.1),
               seawaterDensity=1.025,
               wettedSA=10746.282,
               shipSpeed=seq(1,5,1),
               Cf=c(0.0019,0.0017,0.0016,0.0016,0.0015),
               formFactor=1.275601,
               Ca=0.0003356088,
               serviceMargin=15)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.
}
\seealso{
\code{\link{calcSpeedUnitConversion}}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Holtrop-Mennen Calculations}
\concept{Resistance Calculations}
