% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMWakeFraction.r
\name{calcHMWakeFraction}
\alias{calcHMWakeFraction}
\title{calcHMWakeFraction}
\usage{
calcHMWakeFraction(
  breadth,
  wettedSA,
  maxDraft,
  nProp,
  lwl,
  propDiam,
  formFactor,
  Cf,
  Ca,
  Cbw,
  Cp,
  Cm,
  aftDraft = maxDraft,
  Cstern = 0,
  lcb = 0
)
}
\arguments{
\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{wettedSA}{Wetted hull surface area (vector of numericals, m^2) (see 
\code{\link{calcHMWettedSA}})}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{nProp}{Number of propellers (vector of numericals, see 
\code{\link{calcPropNum}})}

\item{lwl}{Waterline Length (vector of numericals, m) (see \code{\link{calclwl}})}

\item{propDiam}{Propeller diameter (vector of numericals, m) (see 
\code{\link{calcPropDia}})}

\item{formFactor}{Form factor (1+k) (vector of numericals, dimensionless) (see 
\code{\link{calcHMFormFactor}})}

\item{Cf}{Frictional resistance coefficient (vector of numericals, dimensionless)
(see \code{\link{calcCf}})}

\item{Ca}{Incremental hull (roughness) resistance coefficient (vector of 
numericals, dimensionless) (see \code{\link{calcHMCa}})}

\item{Cbw}{Waterline block coefficient (vector of numericals, dimensionless) 
(see \code{\link{calcCbw}})}

\item{Cp}{Prismatic coefficient (vector of numericals, dimensionless) (see 
\code{\link{calcCp}})}

\item{Cm}{Midship section coefficient (vector of numericals, dimensionless) 
(see \code{\link{calcCm}})}

\item{aftDraft}{Aft draft (deviation from actual draft indicates trim) 
(vector of numericals, m)}

\item{Cstern}{Afterbody form coefficient: \itemize{\item V-shaped Hull = -10
        \item U-Shaped Hull =  10
        \item Normal Hull = 0 (default)}
Can supply either a vector of numericals, a single number, or rely on the default}

\item{lcb}{Longitudinal position of center of buoyancy (vector of numericals, 
see \code{\link{calclcb}})}
}
\value{
\code{wakeFraction} (vector of numericals, dimensionless)
}
\description{
Calculate wake fraction (\code{wakeFraction}) (dimensionless)
using the Holtrop & Mennen method.
}
\details{
"The speed of advance of the propeller relative to the water in which it is working is
 lower than the observed speed of the vessel. This difference in speed,
 expressed as a percentage of the ship speed, is known as the wake fraction coefficient".
 \url{https://www.wartsila.com/encyclopedia/term/wake-fraction-coefficient}

Wake fraction is a component of hull efficiency as well as a component of
propeller efficiency.

Actual draft is typically obtained from sources such as AIS messages or ship
records.

Note: In "A Statistical Re-Analysis of Resistance and Propulsion Data", the
authors re-analyze with the inclusion of Series 64 hull forms for a total of
334 models included in the analysis. They suggest an update of the single
screw wake fraction equation but that the original equations should be used
for twin screw ships.

Viscous resistance coefficient: \code{Cv = (1+k) * Cf + Ca}

We are assuming here that \code{1+k = 1+k_1} from \code{\link{calcHMFormFactor}}.

Additionally, we are assuming conventional stern for all single screw ships.
Holtrop & Mennen also include an estimation for wake fraction for single screw
ships with open stern for fast sailing ships, but that is not included here.
}
\examples{
calcHMWakeFraction(c(32.25,32.20),
                   c(10746.28,8669.7),
                   c(13.57,11.49),
                   c(1,1),
                   c(218.75,209.25),
                   c(6.7,7),
                   c(1.27,1.18),
                   c(0.0015,0.0014),
                   c(0.00033,0.00035),
                   c(0.81,0.65),
                   c(0.81,0.67),
                   c(0.99,0.98),
                   c(13.57,11.49))
}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.

Holtrop, J. and Mennen, G. G. J. 1984. "A Statistical Re-Analysis of Resistance
and Propulsion Data'.
}
\seealso{
\itemize{
\item \code{\link{calcPropNum}}
\item \code{\link{calclwl}}
\item \code{\link{calcPropDia}}
\item \code{\link{calcCf}}
\item \code{\link{calcCbw}}
\item \code{\link{calcCp}}
\item \code{\link{calcCm}}
\item \code{\link{calclcb}}
}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcHMWettedSA}()}
}
\concept{Holtrop-Mennen Calculations}
