% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMWaveMakingRes.r
\name{calcHMWaveMakingRes}
\alias{calcHMWaveMakingRes}
\title{calcHMWaveMakingRes}
\usage{
calcHMWaveMakingRes(
  lwl,
  breadth,
  Cp,
  Cwp,
  Cm,
  maxDisplacement,
  maxDraft,
  froudeNum,
  At,
  hb,
  Abt,
  seawaterDensity,
  forwardDraft,
  lcb = 0
)
}
\arguments{
\item{lwl}{Waterline length (vector of numericals, m) (see
\code{\link{calclwl}})}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{Cp}{Prismatic coefficient (vector of numericals, dimensionless) (see
\code{\link{calcCp}})}

\item{Cwp}{Water plane area coefficient (vector of numericals, see
\code{\link{calcCwp}})}

\item{Cm}{Midship section coefficient (vector of numericals, dimensionless)
(see \code{\link{calcCm}})}

\item{maxDisplacement}{Maximum ship displacement (vector of numericals, m^3)}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{froudeNum}{Froude number (vector of numericals, dimensionless) (see
\code{\link{calcFroudeNum}})}

\item{At}{Transom area (vector of numericals, m^2) (see \code{\link{calcAt}})}

\item{hb}{Center of bulb area above keel line (vector of numericals, m) (see
\code{\link{calchb}})}

\item{Abt}{Traverse bulb area (vector of numericals, m^2) (see
\code{\link{calcAbt}})}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can
supply either a vector of numericals, a single number, or rely on the default}

\item{forwardDraft}{Forward draft (deviation from actual draft indicates trim)
(vector of numericals, m)}

\item{lcb}{Longitudinal position of center of buoyancy (vector of numericals,
see \code{\link{calclcb}})}
}
\value{
\code{Rw} (vector of numericals, kN)
}
\description{
Calculate wave resistance (\code{Rw}) (kN) from the Holtrop &
Mennen method.
}
\details{
Note: This calculates resistance, not a coefficient.Therefore, it does not
need to be multiplied by wetted surface area like the frictional resistance
coefficient is.

Note: In "A Statistical Re-Analysis of Resistance and Propulsion Data" the
authors re-analyze with the inclusion of Series 64 hull forms for a total of
334 models included in the analysis. Their original paper insufficiently
modeled high speed craft with Froude number >= 0.55, thus the original wave
making resistance equation is used for \code{froudeNum} < 0.55 and the new
wave making resistance equation is used for \code{froudeNum} >= 0.55. The
authors also include wave making resistance equations in the updated paper for
Froude number < 0.55, but these require more computing power and they mention
that they closely resemble the original equation. Therefore, the original
equation is used for \code{froudeNum} < 0.55.

Extra Info:

\code{c2}: Accounts for the reduction of the wave resistance due to action of
a bulbous bow

\code{c5}: Represents the influence of a transom stern on the wave resistance

\code{i_E}: The angle of the waterline at the bow in the degrees with reference
to center plane but neglecting the local shape at the stem
}
\examples{
 calcHMWaveMakingRes(lwl=c(218.75,209.25),
                     breadth=c(32.25,32.20),
                     Cp=c(0.81,0.67),
                     Cwp=c(0.91,0.84),
                     Cm=c(0.99,0.98),
                     maxDisplacement=c(80097,52382.04),
                     maxDraft=c(13.57,11.49),
                     froudeNum=c(0,0.25),
                     At=c(22.2,18.5),
                     hb=c(5.43,4.6),
                     Abt=c(32.02,27.98),
                     seawaterDensity=1.025,
                     forwardDraft=c(13.57,11.49))
}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.

Holtrop, J. and Mennen, G. G. J. 1984. "A Statistical Re-Analysis of Resistance
and Propulsion Data'.
}
\seealso{
\itemize{
\item \code{\link{calclwl}}
\item \code{\link{calcCp}}
\item \code{\link{calcCwp}}
\item \code{\link{calcCm}}
\item \code{\link{calcFroudeNum}}
\item \code{\link{calcAt}}
\item \code{\link{calchb}}
\item \code{\link{calcAbt}}
\item \code{\link{calclcb}}}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Holtrop-Mennen Calculations}
\concept{Resistance Calculations}
