% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHMWettedSA.r
\name{calcHMWettedSA}
\alias{calcHMWettedSA}
\title{calcHMWettedSA}
\usage{
calcHMWettedSA(lwl, actualDraft, breadth, Cm, Cbw, Cwp, Abt)
}
\arguments{
\item{lwl}{Waterline length (vector of numericals, m) (see
\code{\link{calclwl}})}

\item{actualDraft}{Actual draft (vector of numericals, m)}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{Cm}{Midship section coefficient (vector of numericals, dimensionless)
(see \code{\link{calcCm}})}

\item{Cbw}{Waterline block coefficient (vector of numericals, dimensionless)
(see \code{\link{calcCbw}})}

\item{Cwp}{Water plane area coefficient (vector of numericals, see
\code{\link{calcCwp}})}

\item{Abt}{Traverse bulb area (vector of numericals, m^2) (see
\code{\link{calcAbt}})}
}
\value{
\code{wettedSA} (vector of numericals, m^2)
}
\description{
Calculate hull wetted surface area (\code{wettedSA}) (m^2)
using the Holtrop & Mennen method.
}
\details{
Actual draft is typically obtained from sources such as AIS messages or ship
records.
}
\examples{
calcHMWettedSA(218.75,12.48,32.25,0.99,0.81,0.91,32.02)

}
\references{
Holtrop, J. and Mennen, G. G. J. 1982. "An approximate power prediction
method." International Shipbuilding Progress 29.
}
\seealso{
\itemize{
\item \code{\link{calclwl}}
\item \code{\link{calcCm}}
\item \code{\link{calcCbw}}
\item \code{\link{calcCwp}}
\item \code{\link{calcAbt}}}

Other Holtrop-Mennen Calculations: 
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMFormFactor}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMPwr}()},
\code{\link{calcHMThrustFactor}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWakeFraction}()},
\code{\link{calcHMWaveMakingRes}()}
}
\concept{Holtrop-Mennen Calculations}
