% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHullEff.r
\name{calcHullEff}
\alias{calcHullEff}
\title{calcHullEff}
\usage{
calcHullEff(thrustFactor, wakeFraction)
}
\arguments{
\item{thrustFactor}{Thrust deduction coefficient (vector of numericals, dimensionless)
(see \code{\link{calcHMThrustFactor}} and \code{\link{calcKristThrustFactor}})}

\item{wakeFraction}{Wake fraction coefficient (vector of numericals, dimensionless)
(see \code{\link{calcHMWakeFraction}} and \code{\link{calcKristWakeFrac}})}
}
\value{
\code{hullEff} (vector of numericals, dimensionless)
}
\description{
Calculate the hull efficiency (\code{hullEff}) (dimensionless)
for a specified vessel.
}
\examples{
calcHullEff(c(0.19,0.17), c(0.32,0.29))

}
\seealso{
\itemize{
\item \code{\link{calcHMThrustFactor}}
\item \code{\link{calcKristThrustFactor}}
\item \code{\link{calcHMWakeFraction}}
\item \code{\link{calcKristWakeFrac}}}
}
