% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKristCa.r
\name{calcKristCa}
\alias{calcKristCa}
\title{calcKristCa}
\usage{
calcKristCa(
  shipType,
  actualDisplacement,
  tankerBulkCarrierGCargoShipTypes = c("tanker", "general.cargo", "chemical.tanker",
    "liquified.gas.tanker", "oil.tanker", "other.tanker", "bulk.carrier"),
  containerShipTypes = c("container.ship")
)
}
\arguments{
\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}).
Must align with \code{tankerBulkCarrierGCargoShipTypes} and
\code{containerShipTypes} groupings}

\item{actualDisplacement}{Actual loaded displacement (vector of numericals,
m^3) (see \code{\link{calcActualDisp}})}

\item{tankerBulkCarrierGCargoShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers, bulk carriers and general cargo
vessels (vector of strings)}

\item{containerShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as container ships (vector of strings)}
}
\value{
\code{Ca} (vector of numericals, dimensionless)
}
\description{
Calculate incremental hull (roughness) resistance coefficient
(\code{Ca}) (dimensionless) using the Kristensen method.
}
\details{
Models the effect of realistic hull roughness on resistance, which is not
captured in the frictional and residual resistance coefficients from tank
towing operations.

This method this requires ship types to be grouped. Use the
\code{tankerBulkCarrierGCargoShipTypes}, \code{containerShipTypes} grouping
parameters to provide these ship type groupings. Any ship types not included
in these groupings will be considered as miscellaneous
}
\examples{
calcKristCa(c("bulk.carrier","container.ship"),c(73663.27,216726.45))

}
\references{
Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion
Power of Ships."

\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}
}
\seealso{
\code{\link{calcActualDisp}}

Other Kristensen Calculations: 
\code{\link{calcKristCaa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristPwr}()},
\code{\link{calcKristThrustFactor}()},
\code{\link{calcKristTotalRes}()},
\code{\link{calcKristWakeFrac}()},
\code{\link{calcKristWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Kristensen Calculations}
\concept{Resistance Calculations}
