% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKristCaa.r
\name{calcKristCaa}
\alias{calcKristCaa}
\title{calcKristCaa}
\usage{
calcKristCaa(
  shipType,
  dwt,
  tankerBulkCarrierGCargoShipTypes = c("tanker", "general.cargo", "chemical.tanker",
    "liquified.gas.tanker", "oil.tanker", "other.tanker", "bulk.carrier"),
  containerShipTypes = c("container.ship")
)
}
\arguments{
\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}). 
Must align with \code{tankerBulkCarrierGCargoShipTypes} and
\code{containerShipTypes} groupings}

\item{dwt}{Ship maximum deadweight tonnage (vector of numericals, tonnage)}

\item{tankerBulkCarrierGCargoShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers, bulk carriers and general cargo
vessels (vector of strings)}

\item{containerShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as container ships (vector of strings)}
}
\value{
\code{Caa} (vector of numericals, dimensionless)
}
\description{
Calculate air resistance (\code{Caa}) (dimensionless) using the
Kristensen method.
}
\details{
Models the effect of realistic hull roughness on resistance, which is not
captured in the frictional and residual resistance coefficients from tank
towing operations.

This method this requires ship types to be grouped. Use the
\code{tankerBulkCarrierGCargoShipTypes}, \code{containerShipTypes} grouping
parameters to provide these ship type groupings. Any ship types not included
in these groupings will be considered as miscellaneous vessels.

NOTE: within the container ship section of this calculation, estimations are
made of the ships TEU:\itemize{
\item Feeder: TEU = ((dwt/15.19)^(1/0.9814))
\item Panamax: TEU = ((dwt/28.81)^(1/0.902))
\item PostPanamax: TEU = ((dwt/37)^(1/0.875))
}
These come from: H.O. Kristensen (2016), "Revision of statistical analysis and determination of
regression formulas for main dimensions of container ships based on data from Clarkson".
Kristensen's SHIP DESMO model also uses another dwt/teu relation for post-panamax container ships
with breadth > 49m. This equation has no inverse to map dwt to TEU and thus all post panamaxs are
assumed to have the same dwt/teu relation for all breadths, described above.
}
\examples{
calcKristCaa(c("bulk.carrier","container.ship","other.tanker"),c(70000,191144,20000))
calcKristCaa(c("bulk.carrier","container.ship","other.tanker"),c(70000,191144,20000),
             tankerBulkCarrierGCargoShipTypes=c("bulk.carrier","other.tanker"))

}
\references{
Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion
Power of Ships."

Kristensen, H. O. 2016. "Revision of statistical analysis and determination of
regression formulas for main dimensions of container ships based on data from
Clarkson."

\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}
}
\seealso{
Other Kristensen Calculations: 
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristPwr}()},
\code{\link{calcKristThrustFactor}()},
\code{\link{calcKristTotalRes}()},
\code{\link{calcKristWakeFrac}()},
\code{\link{calcKristWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Kristensen Calculations}
\concept{Resistance Calculations}
