% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKristCr.r
\name{calcKristCr}
\alias{calcKristCr}
\title{calcKristCr}
\usage{
calcKristCr(
  shipType,
  M,
  froudeNum,
  actualDraft,
  breadth,
  Cp,
  tankerBulkCarrierGCargoShipTypes = c("tanker", "general.cargo", "chemical.tanker",
    "liquified.gas.tanker", "oil.tanker", "other.tanker", "bulk.carrier")
)
}
\arguments{
\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}})}

\item{M}{Fineness/slenderness coefficient (vector of numericals,
dimensionless) (see \code{\link{calcShipM}})}

\item{froudeNum}{Froude number (vector of numericals, dimensionless) (see
\code{\link{calcFroudeNum}})}

\item{actualDraft}{Actual draft (vector of numericals, m)}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{Cp}{Prismatic coefficient (vector of numericals, dimensionless) (see
\code{\link{calcCp}})}

\item{tankerBulkCarrierGCargoShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers, bulk carriers and general cargo
vessels (vector of strings)}
}
\value{
\code{Cr} (vector of numericals, dimensionless)
}
\description{
Calculate the residual resistance coefficient (\code{Cr})
(dimensionless) using the Kristensen method with Harvald regressions.
}
\details{
Note: Uses Guldhammer Residual Resistance Calculation and adjusts for bulbous bow, and
breadth/draft ratio.

This method this requires ship types to be grouped. Use the
\code{tankerBulkCarrierGCargoShipTypes} grouping
parameters to provide these ship type groupings. Any ship types not include
this grouping will be considered as miscellaneous vessels.

Actual draft is typically obtained from sources such as AIS messages or ship
records.
}
\examples{
calcKristCr("bulk.carrier", 5.2, 0.1199268, 12.48, 32.25, 0.81)

}
\references{
Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion
Power of Ships."

\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}
}
\seealso{
\itemize{
\item \code{\link{calcShipM}}
\item \code{\link{calcFroudeNum}}
\item \code{\link{calcCp}}
\item \code{\link{calcShipType}}
}

Other Kristensen Calculations: 
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristPwr}()},
\code{\link{calcKristThrustFactor}()},
\code{\link{calcKristTotalRes}()},
\code{\link{calcKristWakeFrac}()},
\code{\link{calcKristWettedSA}()}

Other Resistance Calculations: 
\code{\link{calcCf}()},
\code{\link{calcHMAppendageRes}()},
\code{\link{calcHMBulbousBowRes}()},
\code{\link{calcHMCa}()},
\code{\link{calcHMImmersedTransomRes}()},
\code{\link{calcHMTotalRes}()},
\code{\link{calcHMWaveMakingRes}()},
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristTotalRes}()}
}
\concept{Kristensen Calculations}
\concept{Resistance Calculations}
