% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKristPwr.r
\name{calcKristPwr}
\alias{calcKristPwr}
\title{calcKristPwr}
\usage{
calcKristPwr(
  totalInstalledPwr,
  shipSpeed,
  actualDraft,
  maxDraft,
  shipType,
  lwl,
  breadth,
  maxDisplacement,
  Cb,
  nProp,
  dwt,
  serviceMargin = 15,
  shaftEff = 0.98,
  relRotationEff = 1,
  seawaterTemp = 15,
  seawaterDensity = 1.025,
  pwrUpperBoundPercent = 1,
  pwrLowerBoundPercent = 0.02,
  CmEquationType = "kristensen",
  tankerBulkCarrierShipTypes = c("tanker", "chemical.tanker", "liquified.gas.tanker",
    "oil.tanker", "other.tanker", "bulk.carrier"),
  tugShipTypes = c("service.tug", "tug"),
  roroPaxShipTypes = c("passenger", "ferry.pax", "ferry.ro.pax", "cruise", "cruise.ed",
    "yacht", "ro.ro"),
  gCargoShipTypes = c("general.cargo"),
  containerShipTypes = c("container.ship")
)
}
\arguments{
\item{totalInstalledPwr}{Total installed main engine power (vector of
 numericals, kW) (maximum
continuous rated power)}

\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see
\code{\link{calcSpeedUnitConversion}})}

\item{actualDraft}{Actual draft (vector of numericals, m)}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}),
determined by Stat 5 code: \itemize{
\item"container.ship"
\item"bulk.carrier"
\item"tanker"
\item"general.cargo"
\item"vehicle.carrier"
\item"reefer"
\item"ro.ro"
\item"passenger"
\item"tug"
\item"misc"
}}

\item{lwl}{Waterline length (vector of numericals, m) (see
\code{\link{calclwl}})}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{maxDisplacement}{Maximum ship displacement (vector of numericals, m^3)}

\item{Cb}{Maximum block coefficient (vector of numericals, dimensionless)
(see \code{\link{calcCb}})}

\item{nProp}{Number of propellers (vector of numericals, see
\code{\link{calcPropNum}})}

\item{dwt}{Ship maximum deadweight tonnage (vector of numericals, tonnage)}

\item{serviceMargin}{A service margin to account for weather and sea effects:
\itemize{\item At-sea operations = 15 (Default) \item Coastal operations = 10}
Can supply either a vector of numericals, a single number, or rely on the
default}

\item{shaftEff}{Shaft efficiency (dimensionless). Default = 0.98.
Ratio of power delivered to the propeller and the brake power delivered by
the engine. Can supply either a vector of numericals, a single number, or
rely on the default}

\item{relRotationEff}{Relative rotational efficiency (dimensionless).
Default = 1. Accounts for effect of rotational flow of water around propeller.
Can supply either a vector of numericals, a single number, or rely on the
default}

\item{seawaterTemp}{Sea water temperature. Default = 15 (degrees Celsius). Can
supply either a vector of numericals, a single number, or rely on the default}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can
supply either a vector of numericals, a single number, or rely on the default}

\item{pwrUpperBoundPercent}{Percent of total installed power at which
required power is capped. Default = 1, which indicates required power cannot
exceed \code{totalInstalledPwr}. Can supply either a vector of numericals, a
single number, or rely on the default}

\item{pwrLowerBoundPercent}{Percent of total installed power to act as lower
bound for required power. Default = 0.02, which indicates required power
cannot go below 2\% of \code{totalInstalledPwr}. Can supply either a vector
of numericals, a single number, or rely on the default}

\item{CmEquationType}{Type of equation to estimate the midship section
coefficient (see \code{\link{calcCm}}): \itemize{
\item"kristensen"
\item"benford"
\item"schneekluth"}}

\item{tankerBulkCarrierShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers and bulk carriers.}

\item{tugShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as tugs}

\item{roroPaxShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as RORO and passenger ships}

\item{gCargoShipTypes}{Ship types specified in input \code{shipTypes} to be
modeled as general cargo}

\item{containerShipTypes}{Ship types specified in input \code{shipTypes} to
be modeled as container ships}
}
\value{
power (vector of numericals, kW)
}
\description{
Calculate instantaneous main engine power (kW) using the
Kristensen 2016 updates of Harvald equations.
}
\details{
Primary method from Kristensen (2013). Estimation of some inputs use
methodology from Rakke (2016).

This method this requires ship types to be grouped. Use the
\code{tankerBulkCarrierShipTypes}, \code{tugShipTypes}, \code{roroPaxShipTypes},
\code{gCargoShipTypes}, \code{containerShipTypes} grouping parameters to
provide these ship type groupings. Any ship types not included in these groupings
will be considered as miscellaneous vessels.

Ship speed and actual draft are typically obtained from sources such as AIS
messages or ship records.
}
\examples{
calcKristPwr(
totalInstalledPwr=9363,
shipSpeed=10.8,
actualDraft=12.5,
maxDraft=13.6,
shipType="bulk.carrier",
lwl=218,
breadth=32.25,
maxDisplacement=80097,
Cb=0.8162717,
nProp=1,
dwt=70000,
serviceMargin=15,
shaftEff=0.98,
relRotationEff=1,
seawaterTemp=15,
seawaterDensity=1.025,
CmEquationType="kristensen"
)

}
\references{
Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion
Power of Ships."

Kristensen, H. O. 2016. "Revision of statistical analysis and determination of
regression formulas for main dimensions of container ships based on data from
Clarkson."

\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}

\href{http://hdl.handle.net/11250/2410741}{Rakke, S. G. 2016. "Ship Emissions
Calculation from AIS." NTNU.}
}
\seealso{
\itemize{
\item \code{\link{calclwl}}
\item \code{\link{calcSpeedUnitConversion}}
\item \code{\link{calcCb}}
\item \code{\link{calcPropNum}}
\item \code{\link{calcCm}}
\item \code{\link{calcShipType}}
\item \code{vignette("OverviewOfPowerModels", package="ShipPowerModel")}
\item \code{vignette("Kristensen.Example", package="ShipPowerModel")}
}

Other Kristensen Calculations: 
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristThrustFactor}()},
\code{\link{calcKristTotalRes}()},
\code{\link{calcKristWakeFrac}()},
\code{\link{calcKristWettedSA}()}
}
\concept{Kristensen Calculations}
