% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKristThrustFactor.r
\name{calcKristThrustFactor}
\alias{calcKristThrustFactor}
\title{calcKristThrustFactor}
\usage{
calcKristThrustFactor(
  shipType,
  breadth,
  lwl,
  Cbw,
  propDiam,
  M,
  nProp,
  tankerBulkCarrierShipTypes = c("tanker", "chemical.tanker", "liquified.gas.tanker",
    "oil.tanker", "other.tanker", "bulk.carrier")
)
}
\arguments{
\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}),
determined by Stat 5 code, \itemize{
\item"container.ship"
\item"bulk.carrier"
\item"tanker"
\item"general.cargo"
\item"vehicle.carrier"
\item"reefer"
\item"ro.ro"
\item"passenger"
\item"tug"
\item"misc"
}}

\item{breadth}{Moulded breadth (vector of numericals, m)}

\item{lwl}{Waterline Length (vector of numericals, m) (see
\code{\link{calclwl}})}

\item{Cbw}{Waterline block coefficient (vector of numericals, dimensionless)
(see \code{\link{calcCbw}})}

\item{propDiam}{Propeller diameter (vector of numericals, m) (see
\code{\link{calcPropDia}})}

\item{M}{Fineness/slenderness coefficient (vector of numericals,
dimensionless) (see \code{\link{calcShipM}})}

\item{nProp}{Number of propellers (vector of numericals, see
\code{\link{calcPropNum}})}

\item{tankerBulkCarrierShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers and bulk carriers vessels
(vector of strings)}
}
\value{
\code{thrustFactor} (vector of numericals, dimensionless)
}
\description{
Calculate the thrust deduction factor (\code{thrustFactor})
(dimensionless) using the Kristensen method.
}
\details{
Thrust deduction factor is a component of hull efficiency as well as a
component of propeller efficiency. It describes the increase in resistance on
the hull from water getting sucked back towards the propeller.

This method this requires ship types to be grouped. Use the
\code{tankerBulkCarrierShipTypes} grouping  parameters to provide these ship
type groupings. Any ship types not included in this grouping will be considered
as miscellaneous vessels.
}
\examples{
calcKristThrustFactor(c("bulk.carrier","container.ship"),
                      c(32.25,49),
                      c(218.75,400),
                      c(0.8,0.75),
                      c(6.7,9.6),
                      c(5.2,6.7),
                      c(1,1))

calcKristThrustFactor("bulk.carrier", 32.25, 218.75, 0.8, 6.7,5.2,1)

}
\references{
Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion
Power of Ships."

\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}
}
\seealso{
\itemize{
\item \code{\link{calclwl}}
\item \code{\link{calcCbw}}
\item \code{\link{calcPropDia}}
\item \code{\link{calcShipM}}
\item \code{\link{calcPropNum}}
\item \code{\link{calcShipType}}
}

Other Kristensen Calculations: 
\code{\link{calcKristCaa}()},
\code{\link{calcKristCa}()},
\code{\link{calcKristCr}()},
\code{\link{calcKristPwr}()},
\code{\link{calcKristTotalRes}()},
\code{\link{calcKristWakeFrac}()},
\code{\link{calcKristWettedSA}()}
}
\concept{Kristensen Calculations}
