% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calclcb.r
\name{calclcb}
\alias{calclcb}
\title{calclcb}
\usage{
calclcb(lwl)
}
\arguments{
\item{lwl}{Waterline length (vector of numericals, m) (see \code{\link{calclwl}})}
}
\value{
\code{lcb} (vector of numericals)
}
\description{
Calculate longitudinal position of center of buoyancy
(\code{lcb}).
}
\details{
The longitudinal position of center of buoyancy describes center of buoyancy
forward of half of the waterline length (\code{lwl}) (decimal percentage of
waterline length).
}
\examples{
calclcb(c(218.75,400,262))

}
\references{
\href{http://hdl.handle.net/11250/2410741}{Rakke, S. G. 2016. "Ship Emissions
Calculation from AIS." NTNU.}
}
\seealso{
\code{\link{calclwl}}
}
