% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOpenWaterEff.r
\name{calcOpenWaterEff}
\alias{calcOpenWaterEff}
\title{calcOpenWaterEff}
\usage{
calcOpenWaterEff(
  Rtot,
  thrustFactor,
  nProp,
  wakeFraction,
  propDiam,
  shipSpeed,
  seawaterDensity = 1.025
)
}
\arguments{
\item{Rtot}{Total ship resistance (vector of numericals, kN) (see
\code{\link{calcKristTotalRes}} or \code{\link{calcHMTotalRes}})}

\item{thrustFactor}{Thrust deduction factor (vector of numericals,
dimensionless) (see \code{\link{calcKristThrustFactor}} or
\code{\link{calcHMThrustFactor}})}

\item{nProp}{Number of propellers (vector of numericals, see
\code{\link{calcPropNum}})}

\item{wakeFraction}{Wake fraction coefficient (vector of numericals,
dimensionless) (see \code{\link{calcKristWakeFrac}} or
\code{\link{calcHMWakeFraction}})}

\item{propDiam}{Propeller diameter (vector of numericals, m) (see
\code{\link{calcPropDia}})}

\item{shipSpeed}{Ship actual speed (vector of numericals, m/s) (see
\code{\link{calcSpeedUnitConversion}})}

\item{seawaterDensity}{Sea water density. Default = 1.025 (g/cm^3). Can
supply either a vector of numericals, a single number, or rely on the default}
}
\value{
\code{openWaterEff} (vector of numericals, dimensionless)
}
\description{
Calculate the open water efficiency (\code{openWaterEff})
(dimensionless) using the Kristensen method with Harvald regressions.
}
\examples{
 calcOpenWaterEff(398.487,0.1894947,1,0.3199536,6.7,0,seawaterDensity=1.025)

}
\references{
\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}
}
\seealso{
\itemize{
\item \code{\link{calcKristTotalRes}}
\item \code{\link{calcHMTotalRes}}
\item \code{\link{calcKristThrustFactor}}
\item \code{\link{calcHMThrustFactor}}
\item \code{\link{calcPropNum}}
\item \code{\link{calcKristWakeFrac}}
\item \code{\link{calcHMWakeFraction}}
\item \code{\link{calcPropDia}}
\item \code{\link{calcSpeedUnitConversion}}}
}
