% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPropDia.r
\name{calcPropDia}
\alias{calcPropDia}
\title{calcPropDia}
\usage{
calcPropDia(
  shipType,
  maxDraft,
  tankerBulkCarrierGCargoShipTypes = c("general.cargo", "tanker", "chemical.tanker",
    "liquified.gas.tanker", "oil.tanker", "other.tanker", "bulk.carrier"),
  containerShipTypes = c("container.ship")
)
}
\arguments{
\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}).
Must align with \code{tankerBulkCarrierGCargoShipTypes} and
\code{containerShipTypes} groupings}

\item{maxDraft}{Maximum summer load line draft (vector of numericals, m)}

\item{tankerBulkCarrierGCargoShipTypes}{Ship types specified in input
\code{shipTypes} to be modeled as tankers, bulk carriers and general cargo
vessels (vector of strings)}

\item{containerShipTypes}{Ship types specified in input \code{shipTypes} to
be modeled as container ships (vector of strings)}
}
\value{
\code{propDiam} (vector of numericals, m)
}
\description{
Calculate propeller diameter (\code{propDiam}) (m).
}
\details{
This method this requires ship types to be grouped. Use the
\code{tankerBulkCarrierGCargoShipTypes} and \code{containerShipTypes} grouping
parameters to provide these ship type groupings. Any ship types not included
in these groupings will be considered as miscellaneous vessels.
}
\examples{
calcPropDia(c("bulk.carrier","container.ship"), c(13.6,15.6))
calcPropDia(c("other.tanker","container.ship"), c(13.6,15.6),
            tankerBulkCarrierGCargoShipTypes=c("other.tanker","bulk.carrier"))

}
\references{
Kristensen, H. O. and Lutzen, M. 2013. "Prediction of Resistance and Propulsion
Power of Ships."

\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}
}
