% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPropNum.r
\name{calcPropNum}
\alias{calcPropNum}
\title{calcPropNum}
\usage{
calcPropNum(
  shipType,
  paxTugShipTypes = c("cruise.ed", "passenger", "ferry.pax", "ferry.ro.pax", "cruise",
    "yacht", "tug", "service.tug")
)
}
\arguments{
\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}})}

\item{paxTugShipTypes}{Ship types specified in input \code{shipTypes} to
be modeled as passenger and tug vessels. (vector of strings)}
}
\value{
\code{nProp} (vector of numericals)
}
\description{
Calculate number of propellers (\code{nProp}) by ship type
trends.
}
\details{
Based on a sample of 2,700 ships from Clarksons.

This method this requires ship types to be grouped. Use the
\code{paxTugShipTypes} grouping parameter to provide these ship type
groupings. Any ship types not included in these groupings will be considered
as miscellaneous vessels.
}
\examples{
calcPropNum(c("container.ship","bulk.carrier","tug"))

}
