% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpeedUnitConversion.r
\name{calcSpeedUnitConversion}
\alias{calcSpeedUnitConversion}
\title{calcSpeedUnitConversion}
\usage{
calcSpeedUnitConversion(shipSpeed)
}
\arguments{
\item{shipSpeed}{Ship speed (vector of numericals, knots)}
}
\value{
speed (vector of numericals, m/s)
}
\description{
Convert speed in knots to m/s.
}
\examples{
calcSpeedUnitConversion(seq(10,15,1))

}
