% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTEU.r
\name{calcTEU}
\alias{calcTEU}
\title{calcTEU}
\usage{
calcTEU(shipType, dwt)
}
\arguments{
\item{shipType}{Ship type (vector of strings, see \code{\link{calcShipType}}),
determined by Stat 5 code, \itemize{
\item"container.ship"
\item"bulk.carrier"
\item"tanker"
\item"general.cargo"
\item"vehicle.carrier"
\item"reefer"
\item"ro.ro"
\item"passenger"
\item"tug"
\item"misc"
}}

\item{dwt}{Ship maximum deadweight tonnage (vector of numericals, tonnage)}
}
\value{
TEU (vector of numericals)
}
\description{
Estimate vessel size in twenty-foot equivalent units (TEUs) from
deadweight tonnage (\code{dwt}) for container ships.
}
\examples{
calcTEU(c("bulk.carrier","container.ship"), c(70000,191144))

}
\references{
\href{https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}{Kristensen, H. O.
"Ship-Desmo-Tool." https://gitlab.gbar.dtu.dk/oceanwave3d/Ship-Desmo}
}
