% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{IMOGHG3_AuxLoadFactor}
\alias{IMOGHG3_AuxLoadFactor}
\title{Auxiliary engine loads per operating mode, ship type and sub type using IMO GHG 3 methodology. Also used to define sub types.}
\format{
A data frame with 56 rows and 9 variables:
\describe{
\item{shipType}{Generalizable ship type. Define by \code{calcShipType.r}}
\item{sizeMin}{Minimum size value for each subtype and their given size unit. Used to define subtype.}
\item{sizeMax}{Maximum size value for each subtype and their given size unit. Used to define subtype.}
\item{sizeUnits}{Size units used to define subtype size ranges.}
\item{subType}{Sub type of each ship type which are defined by size ranges.}
\item{Transit}{Auxiliary engine loads (kW) assigned to each vessel by subtype when transiting. Transiting defined by \code{calcOperatingMode.r}}
\item{Maneuvering}{Auxiliary engine loads (kW) assigned to each vessel by subtype when Maneuvering. Maneuvering defined by \code{calcOperatingMode.r}}
\item{Berth}{Auxiliary engine loads (kW) assigned to each vessel by subtype when berth. Berthing defined by \code{calcOperatingMode.r}}
\item{Anchorage}{Auxiliary engine loads (kW) assigned to each vessel by subtype when anchoring. Anchorage defined by \code{calcOperatingMode.r}}
}
}
\usage{
data(IMOGHG3_AuxLoadFactor)
}
\description{
Auxiliary engine loads per operating mode, ship type and sub type using IMO GHG 3 methodology. Also used to define sub types.
}
\details{
IMO GHG 3 uses Starcrest 2014 auxiliary loads from multiple vessel boarding programs(VBP) around the US in addition to FMI international auxiliary load data.

This data is used in \code{calcAuxBoilerLoad.r} \code{calcSubType.r}
}
\examples{
assign("Loads",get(utils::data(IMOGHG3_AuxLoadFactor)))
}
\references{
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
}
\keyword{datasets}
