% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{shipMap}
\alias{shipMap}
\title{Map between individual Clarkson's Vessel_Type and generalizable ship types used in IMO GHG 3 (imoShipType) and Starcrest (starcrestShipType) inventories.}
\format{
A data frame with 341 rows and 3 variables:
\describe{
\item{Vessel_Type}{Specific Clarkson's Vessel_Type. }
\item{imoShipType}{Generalizable ship types alligning with IMO GHG 3 auxiliary and boiler loads.}
\item{starcrestShipType}{Generalizable ship types alligning with starcrest 2016 auxiliary and boiler loads.}
}
}
\usage{
data(shipMap)
}
\description{
Map between individual Clarkson's Vessel_Type and generalizable ship types used in IMO GHG 3 (imoShipType) and Starcrest (starcrestShipType) inventories.
}
\details{
This is used in \code{calcShipType.r}
}
\examples{
data(shipMap)
}
\references{
International Maritime Organization.Third IMO GHG Study 2014 - Final Report. London: International Maritime Organization, 2014.
Starcrest Consulting Group,LLC.Port of Los Angeles Inventory of Air Emissions-2016, 2017.
}
\keyword{datasets}
