% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Starcrest2016_AuxLoadFactor}
\alias{Starcrest2016_AuxLoadFactor}
\title{Auxiliary engine loads per operating mode, ship type and sub type using Starcrest 2016 methodology. Also used to define sub types.}
\format{
A data frame with 34 rows and 9 variables:
\describe{
\item{shipType}{Generalizable ship type. Define by \code{calcShipType.r} using method="starcrest"}
\item{sizeMin}{Minimum size value for each subtype and their given size unit. Used to define subtype.}
\item{sizeMax}{Maximum size value for each subtype and their given size unit. Used to define subtype.}
\item{sizeUnits}{Size units used to define subtype size ranges.}
\item{subType}{Sub type of each ship type which are defined by size ranges.}
\item{Transit}{Auxiliary engine loads (kW) assigned to each vessel by subtype when transiting. Transiting defined by \code{calcOperatingMode.r}}
\item{Maneuvering}{Auxiliary engine loads (kW) assigned to each vessel by subtype when Maneuvering. Maneuvering defined by \code{calcOperatingMode.r}}
\item{Berth}{Auxiliary engine loads (kW) assigned to each vessel by subtype when berth. Berthing defined by \code{calcOperatingMode.r}}
\item{Anchorage}{Auxiliary engine loads (kW) assigned to each vessel by subtype when anchoring. Anchorage defined by \code{calcOperatingMode.r}}
}
}
\usage{
data(Starcrest2016_AuxLoadFactor)
}
\description{
Auxiliary engine loads per operating mode, ship type and sub type using Starcrest 2016 methodology. Also used to define sub types.
}
\details{
This data is used in \code{calcAuxBoilerLoad.r} \code{calcSubType.r}
}
\examples{
assign("Loads",get(utils::data(Starcrest2016_AuxLoadFactor)))
}
\references{
Starcrest Consulting Group,LLC.Port of Los Angeles Inventory of Air Emissions-2016, 2017.
}
\keyword{datasets}
