% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Starcrest2016_BoilerLoadFactor}
\alias{Starcrest2016_BoilerLoadFactor}
\title{Boiler engine loads (kW) per operating mode, ship type and sub type using Starcrest 2016 methodology.}
\format{
A data frame with 34 rows and 9 variables:
\describe{
\item{shipType}{Generalizable ship type. Define by \code{calcShipType.r}}
\item{sizeMin}{Minimum size value for each subtype and their given size unit.}
\item{sizeMax}{Maximum size value for each subtype and their given size unit.}
\item{sizeUnits}{Size units used to define subtype size ranges.}
\item{subType}{Sub type of each ship type which are defined by size ranges.}
\item{Transit}{Boiler engine loads (kW) assigned to each vessel by subtype when transiting. Transiting defined by \code{calcOperatingMode.r}}
\item{Maneuvering}{Boiler engine loads(kW) assigned to each vessel by subtype when Maneuvering. Maneuvering defined by \code{calcOperatingMode.r}}
\item{Berth}{Boiler engine loads (kW) assigned to each vessel by subtype when berth. Berthing defined by \code{calcOperatingMode.r}}
\item{Anchorage}{Boiler engine loads (kW) assigned to each vessel by subtype when anchoring. Anchorage defined by \code{calcOperatingMode.r}}
}
}
\usage{
data(Starcrest2016_BoilerLoadFactor)
}
\description{
Boiler engine loads (kW) per operating mode, ship type and sub type using Starcrest 2016 methodology.
}
\details{
This data is used in \code{calcAuxBoilerLoad.r}
}
\examples{
assign("Loads",get(utils::data(Starcrest2016_BoilerLoadFactor)))
}
\references{
Starcrest Consulting Group,LLC.Port of Los Angeles Inventory of Air Emissions-2016, 2017.
}
\keyword{datasets}
