% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{TestShipsHM}
\alias{TestShipsHM}
\title{Unit test data for Holtrop-Mennen Calculations}
\format{
A data frame with 1 row and 34 variables
}
\usage{
data(TestShipsHM)
}
\description{
Unit test data for Holtrop-Mennen Calculations
}
\details{
This data is used for testing the Holtrop-Mennen power calculations
}
\keyword{test}
\keyword{unit}
