% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{TestShipsKrist}
\alias{TestShipsKrist}
\title{Unit test data for Kristensen Calculations}
\format{
A data frame with 6 row and 29 variables
}
\usage{
data(TestShipsKrist)
}
\description{
Unit test data for Kristensen Calculations
}
\details{
This data is used for testing the Kristensen (SHIP-DESMO) power calculations
}
\keyword{test}
\keyword{unit}
