% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOperatingMode.r
\name{calcOperatingMode}
\alias{calcOperatingMode}
\title{calcOperatingMode}
\usage{
calcOperatingMode(shipSpeed, loadFactor, berthArea, anchorageArea)
}
\arguments{
\item{shipSpeed}{Vessel speed (vector of numericals, knots)}

\item{loadFactor}{Fractional percentage (between 0 and 1) of main engine
required to propel vessel at given speed (vector of numericals) (see
ShipPowerModel library)}

\item{berthArea}{Vector of named berth areas (NA if not at berth) or vector
of booleans (TRUE if at berth, FALSE if not)}

\item{anchorageArea}{Vector of named anchorage areas (NA if not at anchorage) 
or vector of booleans (TRUE if at berth, FALSE if not)}
}
\value{
\code{opMode} (vector of strings). Valid values are: \itemize{
\item "Berth"
\item "Anchorage"
\item "Maneuvering"
\item "Transit"
}
}
\description{
Assigns an operating mode designation according to ship speed and engine load
fraction.
}
\details{
Calculated as a simplification of the methods described in Section 3.8.5 of
the Port Emissions Inventory Guidance.
}
\examples{
calcOperatingMode(shipSpeed = c(0, 2.5, 14, 20), 
                  loadFactor = c(0.01, 0.0, 0.15, 0.75),
                  berthArea = c(TRUE, FALSE, FALSE, FALSE),
                  anchorageArea = c(FALSE, TRUE, FALSE, FALSE)
                  )

}
\references{
\href{https://nepis.epa.gov/Exe/ZyPDF.cgi?Dockey=P10102U0.pdf}{EPA. 2020.
"Port Emissions Inventory Guidance: Methodologies for Estimating
Port-Related and Goods Movement Mobile Source Emissions." Ann Arbor, MI:
Office of Transportation and Air Quality. US Environmental Protection Agency.}
}
\seealso{
ShipPowerModel library
}
